Add more template fields to `DbtBaseOperator` (#786)
This partially addresses #754 via allowing for built-in templating
support for the `DbtBaseOperator`.

I also noticed `--full-refresh` was not documented so I added that in.

## Still missing

Manual run pattern is not documented; the fact that these fields are
templated is not documented. I don't really know where in the docs to
put this. The docs are very API-focused more than narrative-based or
suggestive, and Cosmos's maintainers prefer this style of documentation
so it's hard to find a spot for this. It's possible that that's fine and
we just keep this as a feature for more advanced users who dig into
source code to discover for themselves? 🤷

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>