Make Pydantic a required dependency (#939)
The pull request #736 initially designated Pydantic as an 
optional dependency. However, a subsequent pull request, 
#794 as an [import](https://github.com/astronomer/astronomer-cosmos/pull/794/files#diff-bd3fa47d7a9b96d7bb365f3ba3b60eaf0b20e06a48b28814cf6f6e6fb64d4da6R12),
introduced an implementation that requires Pydantic
to be a mandatory dependency for astronomer-cosmos. 

This pull request addresses the bug encountered while 
running `airflow db init`, as described in #936, by enforcing 
Pydantic as a required dependency.

Closes: #936