Correct stale `root_path` in partial parse file
With the introduction of enabling partial parse in PR #904, upon
testing the implementation, it is observed that the seeds files
were not been able to be located as the partial parse file contained
a stale `root_path` from previous command runs. This PR attempts
to correct the `root_path` in the partial parse file by replacing
it with the needed project directory where the project files are
located.

closes: #937