Enable pre-commit run and fix type-check job (#957)
closes: https://github.com/astronomer/astronomer-cosmos/issues/952

This PR addresses the following issues:

**1. Pre-commit Setup:**
- Pre-commit was not running in the CI pipeline because it was neither
configured for this repository nor were we using the pre-commit command
in CI to run it manually. To fix this:
Installed and added pre-commit to this repository. This allows
pre-commit to automatically run the pre-commit configuration on every
push.

**2. Type Check Job:**
- Previously, the type-check job was running with mypy directly, and the
pre-commit type check command had different configurations. This PR
aligns the configurations.

**3. Type Check Fixes:**
- Ignored the type check for __dataclass_fields__ since mypy seems to
not recognize it.
- Added a check for None for ti.task