Fix `dev/Dockerfile` + Add `uv pip install` for faster build time (#997)
## Fix dev Dockerfile

The Dockerfile was not building because `dbt-postgres==1.8.0` pinned
`psycopg2`.

To resolve this, you need to apt-get install `build-essential` and
`libpq-dev`. I added a `RUN` directive that does that. This is placed
near the top as it makes the most sense for layering purposes.

## Add uv

While I was at it, I added uv: https://github.com/astral-sh/uv tldr:
this makes the Dockerfile build a little faster. This speed is
especially important due to how the Dockerfile will typically be used:
if a user makes a change to the `cosmos/` directory and then wants to
test their code, due to how the layers are stacked, the pip install will
need to be re-run and resolved on over 100 Python packages on each code
change.