Improve error logging in DbtLocalBaseOperator (#1004)
Improve error logging when the `dbt` command returns a non-zero exit
code. Instead of raising an `AirflowException` with the full output, log
the output using the logger and then raise the exception with a concise
error message. This makes the dbt output more readable and not in a
single line as AirflowException logs message in a single line, and it can 
get very long. 

closes #1003 
---------

Co-authored-by: David Steinar Asgrimsson <david-steinar.asgrimsson@kommuninvest.se>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>