Optimise performance: only run dbt deps when there are dependencies
As of Cosmos 1.4, Cosmos will attempt to run dbt deps even if there are no dependencies.yml or packages.yml in the dbt project directory.
This causes an unecessary overhead in creating subprocesses, without any benefit.

This problem was originally spotted by Algirdas, who created a pull request proposing a solution to the problem:
https://github.com/astronomer/astronomer-cosmos/pull/893/commits

Since then, the PR became stale, but the problem remains relevant.

This PR proposes a different implementation to solve the same problem.

Co-authored-by: AlgirdasDubickas <123624084+AlgirdasDubickas@users.noreply.github.com>