Bring back `dataset` as a required field for BigQuery profile (#1033)
In PR #1017, we attempted to remove `dataset` from the required fields
list for the BigQuery profile. However, we realised that this is failing
BiqQuery dbt operations as it indeed is a required field. Hence, bring
back the same as a required field. This is also necessary for building
the mock profile where we construct the profile by taking in
consideration only the required fields.

Closes: #1031
(cherry picked from commit 803776abbeaf1f1f8ea739f7e73c324e7e8edebc)