[Bug] Accidentally attaching adapters to root logger leads to increased logging costs (#1047)
Closes #1048 

After having upgraded to `astronomer-cosmos==1.4.3` we noticed an
unusual 2x increase of the cost of the logs in our cloud provider. Quick
investigation showed that the logs from cosmos logger were duplicated
multiple times.
**Example:**
```bash 
[2024-06-17 12:18:48,259] {{manager.py:165}} INFO - Launched DagFileProcessorManager with pid: 656
2024-06-17T12:18:48.260107549Z [2024-06-17 12:18:48,259] {manager.py:165} INFO - (astronomer-cosmos) - Launched DagFileProcessorManager with pid: 656
2024-06-17T12:18:48.260225215Z [2024-06-17 12:18:48,259] {manager.py:165} INFO - (astronomer-cosmos) - Launched DagFileProcessorManager with pid: 656
2024-06-17T12:18:48.260345799Z [2024-06-17 12:18:48,259] {manager.py:165} INFO - (astronomer-cosmos) - Launched DagFileProcessorManager with pid: 656
```

After reproducing the problem locally I've found the highest version
that seemed to not have the problem, which was `1.3.2`. However, the
problem repeated itself when `cosmos` was imported in the dag.
Afterwards, I looked at the changed files and saw that airflow plugin
was added in the 1.4.0 version. It explains why the issue reproduces in
all the version, but is hidden at first in 1.3.2
```toml
[project.entry-points."airflow.plugins"]
cosmos = "cosmos.plugin:CosmosPlugin"
```

Following this discovery i looked through the source code and saw that
in some of the lines the logger was initialized though function
`get_logger` without the name argument. If you call
`logging.getLogger(None)`, you get a root logger. Hence, cosmos adapter
got attached to root logger multiple times in these places.

```python 
def get_logger(name: str) -> logging.Logger:
    logger = logging.getLogger(None) # Now it is a root logger
    ....
    logger.addHandler(handler) # now we have added astronomer logger to root logger, which is not intented
```

## What was done

- I made the `name` argument in `get_logger` required, otherwise it
would be easy to forget to add it when a new file with logger call is
created.
- Added `LOGGER_NAME_TEMPLATE` so that there are no collisions with the
adapters attaching to wrong loggers in the future. The name is not used
in the actual format of the log message, so it does not matter much.

The previous tests did not catch this corner case, but when the argument
is required, there is no need to check it