Change example DAGs to use example_conn as opposed to airflow_db
It can be quite misleading to be using as an example connection the Airflow metadata datase connection.
In our test base, in particular, when we execute Airflow, it's creating a sqlite database by default, and the environment variable value is mismatched with the actual Airflow db. This change decouples both.