Fix getting temporary AWS credentials with assume_role (#1081)
When Airflow is getting temporary AWS credentials by assuming role with
`role_arn` as only `Connection` parameter, this cause task to fail due
to missing credentials. This is due to the latest changes related to
profile caching. The `env_vars` are accessed before `profile` which, in
this case, means required values are not populated yet.