Fix issue 'No such file or directory'
Two Astronomer customers reported issues while trying to upgrade to Cosmos 1.5:

```
No such file or directory "/usr/local/aiflow/dags/dbt/dbt_venv/bin/python"
```

The issue is that their dbt project folder had a symbolic link to a file that no longer
existed. When Cosmos attempted to iterate over that directory to create a hash representing
the cache version, it could not open the file referenced by the symbolic link.

This PR reproduces the issue and fixes the behaviour. Users are also advised to
check if there are any broken sym links on their dbt folder and delete them.

Closes #1096