Use `dbt ls` as the default parser when profile_config is provided (#1101)
The `load_via_custom_parser` method is not very effective 
when it comes to filtering. Currently, if the ExecutionMode 
is not Local, we always use load_via_custom_parser to parse 
the dbt project. However, I believe it's beneficial to utilize DBT LS 
whenever it's available. This PR removes the check if the 
execution mode is ExecutionMode.LOCAL before `load_via_dbt_ls`

Related: https://github.com/astronomer/astronomer-cosmos/issues/943