Cache package-lock.yml file (#1086)
This PR aims to cache the package-lock.yml in `cache_dir/dbt_project`

Since dbt version 1.7.0, executing the dbt deps command results in the
generation of a package-lock.yml file. This file pins the dependencies
and their versions for the dbt project. dbt uses this file to install
packages, ensuring predictable and consistent package installations
across environments.

- This feature is enabled only if the user checks in package-lock.yml in
their dbt project. Also, I'm assuming if `package-lock.yml` their
dbt-core version is >= 1.7.0 since this feature is available for only
dbt >= 1.7.0
- package-lock.yml also contains the sha1_hash of the packages. This is
used to check if the cached package-lock.yml is outdated or not in this
PR
- The cached `package-lock.yml` is finally copied from from cached path
to the tmp project and used
- To update dependencies or versions, it is expected that the user will
manually update their package-lock.yml in the dbt project using the dbt
deps command.


closes: https://github.com/astronomer/astronomer-cosmos/issues/930