Simplify logging (#1108)
This PR addresses #906 and fixes issues in the Cosmos logging once and
for all.\*

> \* Actually, there is another issue with logs being polluting with
warnings about "source" nodes in 1.5.0, but that is a separate matter!

I have a long explanation of how the `logging` module in Python works,
and the sort of idioms it expects of end users of the module, here:
https://github.com/astronomer/astronomer-cosmos/issues/906#issuecomment-2134150103

The choices I made, explained:

- Although I don't know that I entirely agree with adding
`(astronomer-cosmos)` to all the logs, clearly at least one user, and
possibly many more, want it, and I don't believe we should remove it.
The objective of this PR was therefore to preserve the feature while
future-proofing against future issues.
- Why I can't say I'm a fan of it: It seems that adding
`(astronomer-cosmos)` to logs is a symptom of other problems with the
Cosmos library, specifically how it impacts performance when users do
not set it up effectively. And the prefix was added as a way to assist
people in diagnosing these issues. I think ultimately we want to move
away from this. Other components of the Airflow ecosystem do not feel
compelled to do things like this. Also, the module path is something
that can be handled in the `log_format` if users really want it.
- How I future-proofed: As per the long post I link above, basically the
issue is that there should not be tons of StreamHandlers being created.
The proper and typical use of the logging module, with few exceptions,
is to allow for logs to propagate upwards to a root logger. The reason
the Cosmos logs presented issues for so long was because it deviated a
lot from this.
- I think default behavior being the "least astonishing" means making no
modifications to the base logging behavior whatsoever. This is also less
likely to morph into future issues if any further changes are made to
the custom logging.
- One thing I never mentioned: I found it odd that by default Cosmos did
not "work out of the box" and that, despite using Astronomer's own
Airflow platform (!), I had to set a config option that made Cosmos
logging not be a nightmare (i.e. set `propagate_logs` = false). Previous
logs referenced the Celery Executor as having issues, even though this
is one of 2 of the most popular production ways to run Airflow.
Something like this should just work out of the box for a majority of
users!
- For task execution, Cosmos should make use of the more
Airflow-idiomatic `LoggingMixin` class whenever appropriate. This can
also be used in scheduler / webserver related logging contexts but I
think it is less out-of-place there to use globally scoped loggers.
- These will not use the `get_logger()` implementation. That is
intentional and probably desirable. These logs do not need to be
"enriched" because they are isolated in the task execution logs.

Oh also, I fixed an issue in the `project.entry_points` in the
`pyproject.toml` while I was at it.

## Breaking Change?

- Removes `propagate_logging` conf option, although removing this will
not break users' builds. There is now a `rich_logging` conf option
instead, which by default is disabled.