Resolve CI failures for Apache Airflow 2.7 jobs (#1182)
## Description

This is an effort to fix the currently broken CI jobs. The CI pipeline
was running smoothly until the Cosmos 1.6.0 release (#1103). However,
after the recent release of Apache Airflow providers, we’ve been
encountering deep resolution errors specifically with Apache Airflow 2.7
jobs.

To address this:

1. For Airflow 2.7 tests, we’re pre-installing the providers using
Airflow 2.7 constraints.
2. We’re also upgrading certain jobs, like type-checks and dbt 1.5.4
tests, from Airflow 2.7 to Airflow 2.8, as it seems beneficial to move
from 2.7 to 2.8.

Despite pre-installing the providers, we’re noticing that on Python 3.8
and 3.9 with Apache Airflow 2.7, Cosmos still attempts to sync
dependencies and continues to encounter deep resolution errors.
Therefore, we’re specifically excluding those combinations from running
in Unit and Integration test jobs. However, we have Python 3.10 and 3.11
jobs enabled and running fine for Airflow 2.7 in our CI.

## Related Issue(s)
related: #1180