Fix CI issues - Upgrade actions/upload-artifact & actions/download-artifact to v4 and set min version for packaging (#1208)
I am observing that our CI
[jobs](https://github.com/astronomer/astronomer-cosmos/actions/runs/10920015707/job/30308959677?pr=1207)
are failing complaining that we're using deprecated version (v2) for
`actions/upload-artifact` & `actions/download-artifact`. It is weird
that the jobs directly are failed for using deprecating versions, but so
is the case that's happening. Hence, to get the CI working I'm upgrading
the versions to use v4.

Additionally, we are observing failures for packaging version as
discussed in
https://github.com/astronomer/astronomer-cosmos/pull/1201#discussion_r1754867454.
Hence, also set the min packaging version to `22.0` as per the
discussion in that thread

closes: https://github.com/astronomer/astronomer-cosmos/issues/1209

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>