Remove PR branch added for testing a change in CI in #1224 (#1233)
We temporarily added a PR (#1224) branch to the CI test.yml to test a
new example DAG, along with the necessary environment variables for the
DAG. However, I missed adding a review comment to remove the branch
after testing. This PR removes the temporarily added branch from the CI
configuration.