Fix displaying dbt docs as menu item in Astro (#1280)
As of Cosmos 1.7.0, Astro customers who have the role of "Organization
Member" or "Workspace Operator" are not able to see the menu item "dbt
Docs" under "Browser". They could access the `/cosmos/dbt_docs` view
directly, though. This PR solves this issue.

Closes: #1131 

# About the solution

This PR follows the Astro
[documentation](https://www.astronomer.io/docs/learn/using-airflow-plugins#flask-blueprints-and-appbuilder-views)
recommendation:
> If you want to use custom menu items in an Airflow environment hosted
on Astro, you must give your plugin the necessary permissions. To do
this, use the @has_access decorator to give your BaseView method
ACTION_CAN_ACCESS_MENU permissions.

By adding:
```
    @has_access([(permissions.ACTION_CAN_ACCESS_MENU, "Custom Menu")]) 
```

# How the fix was tested

In addition to the unit test that checks if all Cosmos plugin endpoints
have the expected permissions, I also validated this change with a user
with the "Organization Member" role. I confirmed that they could see the
menu item and that it redirects to the correct page:

![Screenshot 2024-10-24 at 10 24
03](https://github.com/user-attachments/assets/ea835086-36b6-438f-9e2c-9696cdda8b81)